+++ src/components/Form.jsx
@@ -1,54 +1,45 @@
-import { cloneElement } from 'react';
 import Button from './Button';
 
-function Inputs({ inputs, button, onChangeHandler }) {
-  return inputs.map((input) => {
-    return (
-      <div key={input.id} className="form-item">
-        <label htmlFor={input.name}>{input.label}</label>
-        <input
-          id={input.name}
-          data-id={input.id}
-          name={input.name}
-          type={input.type}
-          value={input.value}
-          onChange={onChangeHandler}
-        />
-        {button && <Button {...{ ...button, id: input.id }} />}
-        {/* {button} */}
-      </div>
-    );
-  });
+function Input({ input, onChangeHandler }) {
+  return (
+    <>
+      <label htmlFor={input.name}>{input.label}</label>
+      <input
+        id={input.name}
+        data-id={input.id}
+        name={input.name}
+        type={input.type}
+        value={input.value}
+        onChange={onChangeHandler}
+      />
+      {input.button && <Button {...input.button} />}
+    </>
+  );
 }
 
-export default function Form({ props }) {
-  // const newElement = children && cloneElement(children[0], { foo: '0' });
+function FormItemList({ inputs, button }) {
+  console.log(inputs);
   return (
-    <form>
-      {/* {inputs.map((input) => {
+    <ul>
+      {inputs.map((input) => {
         return (
-          <div className="form-item" key={input.id}>
-            <label htmlFor={input.name}>{input.label}:</label>
-            <input
-              id={input.name}
-              name={input.name}
-              type={input.type}
-              value={input.value}
-              onChange={onChangeHandler}
-            ></input>
-          </div>
+          <li key={input.id}>
+            <Input input={input} onChangeHandler={input.onChangeHandler} />
+          </li>
         );
-      })} */}
+      })}
+      {button && <Button {...button} />}
+    </ul>
+  );
+}
 
-      {props.map((prop) => {
-        return (
-          <Inputs
-            key={prop.id}
-            inputs={prop.inputs}
-            button={prop.button}
-            onChangeHandler={prop.onChangeHandler}
-          />
-        );
+export default function Form({ props, className }) {
+  console.log(props);
+  console.log(Object.entries(props));
+  return (
+    <form className={className}>
+      {Object.entries(props).map(([key, set]) => {
+        return <FormItemList key={key} inputs={set.inputs} button={set.button} />;
       })}
     </form>
   );