+++ src/components/Skills.jsx
@@ -1,10 +1,12 @@
 import { useState } from 'react';
 import Form from './Form';
-import setInputProps from '../utilities/setInputProps';
+import createInputsProps from '../utilities/createInputsProps';
 import setInputEventHandler from '../utilities/setInputEventHandler';
 import '../styles/skills.css';
+// https://codesandbox.io/p/sandbox/react-dev-64n8l5?file=%2Fsrc%2FTaskList.js&utm_medium=sandpack
+// https://www.w3.org/WAI/tutorials/forms/labels/
+let nextId = 3;
 
-let nextId = 2;
 export default function Skills({ isEditing }) {
   const [skillsData, setSkillsData] = useState({
     skill: {
@@ -12,92 +14,107 @@ export default function Skills({ isEditing }) {
       value: '',
     },
     skills: [
-      { id: 0, value: 'test' },
-      { id: 1, value: 'test again' },
+      { id: 1, value: 'test' },
+      { id: 2, value: 'test again' },
     ],
   });
 
-  const onChangeHandler = setInputEventHandler(skillsData, setSkillsData, true);
-  const onChangeHandlerUpdate = setInputEventHandler(skillsData, (id, value) => {
-    const newSkills = Object.values(skillsData.skills).map((skill) => {
-      return skill.id === +id ? { id: +id, value } : skill;
-    });
+  const onChangeHandler = (e) => {
+    console.log('onChangeHandler firing!');
+    console.log(e.currentTarget);
+  };
 
-    setSkillsData({
-      ...skillsData,
-      skills: newSkills,
-    });
-  });
+  const changeSkillHandler = (e) => {
+    console.log(`changeSkillHandler is firing!`);
+    console.log(e.currentTarget);
+  };
 
-  const addSkillHandler = () => {
-    setSkillsData({
-      skill: {
-        ...skillsData.skill,
-        value: '',
-      },
-      skills: [...skillsData.skills, { id: nextId++, value: skillsData.skill.value }],
-    });
+  const addSkillHandler = (e) => {
+    console.log(`addSkillHandler firing!`);
   };
 
   const deleteSkillHandler = (e) => {
-    setSkillsData({
-      ...skillsData,
-      skills: [...skillsData.skills.filter((skill) => skill.id !== +e.currentTarget.dataset.id)],
-    });
+    console.log(`deleteSkillHandler`);
   };
 
-  const propsForInputs = isEditing && setInputProps({ skill: { ...skillsData.skill } });
-  const propsForSkills =
-    isEditing &&
-    setInputProps(
-      Object.values(skillsData.skills).reduce((accumulator, currentValue) => {
-        return {
-          ...accumulator,
-          [`skill${currentValue.id}`]: { ...currentValue },
-        };
-      }, {}),
-    );
+  const formProps = isEditing && {
+    default: {
+      inputs: [
+        ...createInputsProps(
+          { skill: { ...skillsData.skill } },
+          {
+            onChangeHandler,
+            button: {
+              text: 'Add',
+              className: 'btn skill-add',
+              clickHandler: addSkillHandler,
+            },
+          },
+        ),
+      ],
+    },
+
+    set: {
+      inputs: [
+        ...createInputsProps(
+          { ...skillsData.skills },
+          {
+            className: 'visually-hidden',
+            name: 'skill',
+            onChangeHandler: changeSkillHandler,
+            button: {
+              text: 'Delete',
+              className: 'btn skill-delete',
+              clickHandler: deleteSkillHandler,
+            },
+          },
+        ),
+      ],
+    },
+  };
 
   return (
     <section className="skills">
       <div>
         <h2>Skills</h2>
-        {isEditing ? (
-          <Form
-            props={[
-              {
-                id: 0,
-                inputs: propsForInputs,
-                onChangeHandler: onChangeHandler,
-                button: { text: 'Add', className: 'skill-add', clickHandler: addSkillHandler },
-              },
-              {
-                id: 1,
-                inputs: propsForSkills,
-                onChangeHandler: onChangeHandlerUpdate,
-                button: {
-                  text: 'Delete',
-                  className: 'skill-delete',
-                  clickHandler: deleteSkillHandler,
-                },
-              },
-            ]}
-          ></Form>
-        ) : (
-          // <List arr={skillsData.skills} />
-          <div>Hello world!</div>
-        )}
+        {isEditing ? <Form className="form_skills" props={formProps} /> : <div>Loading</div>}
       </div>
     </section>
   );
 }
 
-function List({ arr }) {
-  return (
-    <ul>
-      {arr.map((item) => {
-        return <li key={item.id}>{item.value}</li>;
-      })}
-    </ul>
-  );
-}
+/*
+<form>
+  <ul>
+    <li class='form-item'>
+      <label>Skill:</label>
+      <input type='text'></input>
+    </li>
+    <button>Add</button>
+  </ul>
+
+  <ul>
+    <li class='form-item'>
+      <label>Skill0:</label>
+      <input type='text'></input>
+    </li>
+    <button>Delete</button>
+  </ul>
+
+  <ul>
+    <li class='form-item'>
+      <label>Skill1:</label>
+      <input type='text'></input>
+      <button>Delete</button>
+    </li>
+  <ul>
+
+  <ul>
+    <li class='form-item'>
+      <label>Skill2:</label>
+      <input type='text'></input>
+      <button>Delete</button>
+    </li>
+  </ul>
+</form>
+*/